// pw-cp-frontend/src/app/core/models/data-product-model.ts
export interface DataProduct {
    id: number;
    name: string;
}

export interface DataMode {
    id: number;
    name: string;
}

export interface DataCurrency {
    id: number;
    name: string;
    symbol: string;
}

export interface DataProducts {
    productId: number;
    productName: string;
    modeId: number;
    modeName: string;
}