// pw-cp-frontend/src/app/core/models/data-client-model.ts
export interface DataDocumentType {
    id: number;
    documentName: any;
}
export interface DataClient {
    clientBank: any;
    documentId: number;
    numberOfDocument: any;
    documentName: string;
    commercialProfileId: number;
    creationDate: any;
    id: number;
    linkage: any;
    fullName: string;
    segmentId: number;
    segmentName: string;
    score: string;
    amount: number;
    rate: number;
    customerCode: any;
    comercialProfileEmpresarioId: number;
    comercialProfileSuperFacilId: number;
}